requirements.txt
click==7.1.1
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.0
Flask-RESTful==0.3.7
Flask-Marshmallow==0.10.1
flask-cors==3.0.10 